create pdf file on the fly

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check')
 
 const Product = require('../models/product');
@@ -164,6 +166,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl)
                 product.imageUrl = image.path
             }
             return product.save().then(result => {
@@ -201,8 +204,13 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
     // Product.deleteById(prodId)
-    Product.deleteOne({ _id: prodId, userId: req.user_id })
-        .then(() => {
+    Product.findById(prodId).then(product => {
+            if (!product) {
+                return next(new Error('product not found!'))
+            }
+            fileHelper.deleteFile(product.imageUrl)
+            return Product.deleteOne({ _id: prodId, userId: req.user_id })
+        }).then(() => {
             console.log('destroyed product')
             res.redirect('/admin/products')
         })
